my current implementation of tracy instrumentation * main loop frame, and physics frame marking * various sub main loop frame marking. * input event zone with description * gdscript function calls with source location and line number
Changed output of stringify to return String so that it can be concatenated with other strings.
Scoped out lots more interesting parts to me, also the gdscript VM to fill out zoning when calling gdscript functions better.

After reading a bit more, having the trace only enable on debug is exactly the opposite of the desired outcome, so I re-arranged everything so that it is separated from DEBUG_ENABLED, and can run in release mode.

tracy was behaving badly, it was because i had removed the defines from the other targets causing the destructor to fail.
https://github.com/wolfpld/tracy/issues/610

Trying to get threads working in godot, but its proving to be weird.